Made a lot of errors go away
