+++ project/__init__.py
@@ -8,6 +8,18 @@ from flask_login import LoginManager
 db = SQLAlchemy()
 jrnl_db = SQLAlchemy() # Does this work? I guess I'll find out
 
+# Is this where it needs to be? IDK
+class journal(db.Model):
+    id = jrnl_db.Column('entry_id', jrnl_db.Integer, primary_key = True)
+    time = jrnl_db.Column(jrnl_db.String(50))
+    title = jrnl_db.Column(jrnl_db.String(100))
+    entry = jrnl_db.Column(jrnl_db.String(200))
+
+    def __init__(self, time, title, entry):
+        self.time = time
+        self.title = title
+        self.entry = entry
+
 def create_app():
     app = Flask(__name__)
